cleanup telemetry
